+++ src/sendData.h
@@ -1,7 +1,3 @@
-//
-// Created by John Beeler on 2/18/19.
-//
-
 #ifndef TILTBRIDGE_SENDDATA_H
 #define TILTBRIDGE_SENDDATA_H
 
@@ -15,10 +11,6 @@
 #include <ArduinoJson.h>
 #include <Ticker.h>
 
-#if (ARDUINO_LOG_LEVEL == 6)
-#include <StreamUtils.h>
-#endif
-
 #include <WiFi.h>
 #include <MQTT.h>
 #include <WiFiMulti.h>
@@ -29,9 +21,9 @@
 #include <LCBUrl.h>
 #include <ArduinoLog.h>
 
-#define GSCRIPTS_DELAY (10 * 60 * 1000)       // 10 minute delay between pushes to Google Sheets directly
-#define BREWERS_FRIEND_DELAY (15 * 60 * 1000) // 15 minute delay between pushes to Brewer's Friend
-#define BREWFATHER_DELAY (15 * 60 * 1000)     // 15 minute delay between pushes to Brewfather
+#define GSCRIPTS_DELAY (10 * 60)       // 10 minute delay between pushes to Google Sheets directly
+#define BREWERS_FRIEND_DELAY (15 * 60) // 15 minute delay between pushes to Brewer's Friend
+#define BREWFATHER_DELAY (15 * 60)     // 15 minute delay between pushes to Brewfather
 
 #define BREWFATHER_MIN_KEY_LENGTH 5
 #define BREWERS_FRIEND_MIN_KEY_LENGTH 12
@@ -42,7 +34,7 @@
 #define GSCRIPTS_MIN_EMAIL_LENGTH 7
 #define GSHEETS_JSON 512
 
-// This is me being simplifying the reuse of code. The formats for Brewer's
+// This is me being lazy and simplifying the reuse of code. The formats for Brewer's
 // Friend and Brewfather are basically the same so I'm combining them together
 // in one function
 #define BF_MEANS_BREWFATHER 1